Fix bug: remove blank target from regular links except for OMALS vocabulary url
This bug was introduce by the pull request #122: Customized submission
form for OMALS collection. It intended to lead the link of “CAB
Agricultural and Environmental Thesaurus” to a blank page. But it
influenced other regular links like “community” and “Role Assign”
because of the change of elements.xsl. Here a special name is assigned
to “CAB Agricultural and Environmental Thesaurus” link, it will avoid
such side affects.

Resolved: bugs introduced by pull request #122: Customized submission
form for OMALS collection.