Added files via upload
Made several changes to the default Symplectic to DSpace crosswalk; I specifically added several fields that were not in the original XSL, as well as changed some of Symplectic's default mappings. I'm not entirely sure if all applicable fields are included; there are 4 schemas (publications, grants, teaching, and another), so more information might be needed. Creative Commons import remains a mystery (I've asked Adrien and Andrew to comment on this). Let the testing begin!