Mapped acceptance-date to dc.date.provenance
Mapped the Elements metadata field "acceptance-date" to the Dublin Core field "dc.description.provenance". "acceptance-date" refers to the date that an item was accepted in Elements. "acceptance-date" was initially mapped to "dc.date.accessioned", but "dc.date.accessioned" technically refers to the date an item was added to DSpace, NOT Elements, so it wouldn't be accurate to map this field to "dc.date.accessioned."