Added revised Sherpa/Romeo green message
Added "you may deposit a draft, submitted, or peer-reviewed version of your work. If the publisher allows, you may also deposit the final published version of your work, " to the S/R green message. Not sure what to do with the second sentence (maybe we could remove "of your work"?). Will test and see if a line break can be inserted. I know we want to be pithy!