Fix broken links on error pages
This is a fix for GitHub issue #264
Incorrect links and non-functioning search box on error page

This adjusts login, home, and discovery links to work on pages
without the complete context, including DSpace error pages.

It's unlikely that we will ever add additional path components to the
VTechWorks urls, for example, changing https://vtechworks.lib.vt.edu/ to
https://vtechworks.lib.vt.edu/papers/, but if we do we will need to revisit
this issue.