Initial port of vtmirage2 theme to DSpace 6
This work is is incomplete. For example, the logo in the header size
is smaller than the current DSpace 5 version, and the i18n label for
number of downloads is missing.

However, the header and footer are both slated to be replaced with
the new Virginia Tech standard header and footer.