Create cfg section for VTechWorks customizations
This adds a section to local.cfg file for VTechWorks
customizations that are common across all of our
systems. I think it makes sense to put even common
configurations in this file now, to simplify future
major upgrades by making it simple to see what we've
modified from default settings.

The dspace.cfg file and other configuration files
should now be used as a catalog of possibilities,
as well as a place to see the default setting for
a configuration.