Fix usage-statistics configuration names
As of DSpace 6, DSpace configuration parameters that are part
of modules must be prepended with the module name in the configuration
file. This is true when the module parameters are configured within a
specific module configuration file, or when the parameters are included
in the local.cfg file, which we are using for VTechWorks configuration.

I've added "usage-statistics" to the front of the parameters that
Philippe added in order to configure the number of results on the
statistics page.

Instead of squashing this commit down into the previous one that
added these parameters, I'll leave it separate as a reminder
while updating the rest of our configuration. I wouldn't want anyone
to waste an hour trying to figure out why the parameters are invisible
to VTechWorks!