Stop logo from being forced to 22px
This commit was originally 732ef1fce6b03 from Amber.
Resolved issue #22, and now resolves issue #568 for DSpace 6.

Commit was originally excluded as we were planning to switch to a new theme
with a different logo.

This change wasn't working when the file was copied to /dspace/modules
area as an overlay and then changed. Not 100% clear that this wasn't
due to browser caching and/or build issues, so this may be revisited
in the next theme (if necessary).