Fix incorrect content disposition threshold
The xmlui content disposition threshold was incorrectly set to -1
instead of 0. In some browsers, this caused most files to download
to the download directory instead of displaying inline. This also
broke certain video playback features, such as scan-ahead, and
in some cases required that the entire video file download before
playing.