Add overlay for Christian Scheible's NPE fix for OAI indexer
This fixes fatal errors while indexing the OAI index. The change has
already been applied to this overlay. The change is in commit,
94c3166f2, which maps to commit  64e269568 in Christian's repository.

This commit was applied first to the main codebase to use Christian's
original commit message and credit, and also copied to the modules
directory to work with our current build process.

If this fix is included in the next DSpace release, we should remove
the overlay from VTechWorks.