Update removal policy link in footer
The link now goes directly to the
"Requesting that Material be Amended or Removed"
section of our policies page.

The anchor names on our policy page  appear to have been generated
automatically; I suspect that this link in the footer  may need updating
occassionally if the anchors in case the anchor names are automatically
regenerated.